+++ lib/computable_subscriber.dart
@@ -25,15 +25,25 @@ class ComputableSubscriber<T> with ComputableMixin<T> implements Computable<T> {
 
   /// Subscribes the subscriber [Computable] to the provided source [Computable].
   StreamSubscription<S> subscribe<S>(
-    Computable<S?> computable, [
-    void Function(S data)? listener,
-  ]) {
+    Computable<S?> computable,
+    void Function(S data)? listener, {
+    bool disposeOnDone = false,
+  }) {
+    final initialValue = computable.get();
+
+    if (initialValue is S) {
+      listener?.call(initialValue);
+    }
+
     final subscription = computable
         .stream()
+        .skip(1)
         .where((value) => value is S)
         .cast<S>()
         .listen((value) {
       listener?.call(value);
+    }, onDone: () {
+      dispose();
     });
     _subscriptions.add(subscription);
     return subscription;
@@ -41,37 +51,34 @@ class ComputableSubscriber<T> with ComputableMixin<T> implements Computable<T> {
 
   /// Subscribes the subscriber [Computable] to the provided source [Stream].
   StreamSubscription<S> subscribeStream<S>(
-    Stream<S> stream, [
-    void Function(S data)? listener,
-  ]) {
+    Stream<S> stream,
+    void Function(S data)? listener, {
+    bool disposeOnDone = false,
+  }) {
     return subscribe(Computable.fromStream<S?>(stream), (value) {
       listener?.call(value);
-    });
+    }, disposeOnDone: disposeOnDone);
   }
 
   /// Subscribes the subscriber [Computable] to the provided source [Future].
-  StreamSubscription<S> subscribeFuture<S>(Future<S> future,
-      [void Function(S data)? listener]) {
+  StreamSubscription<S> subscribeFuture<S>(
+    Future<S> future,
+    void Function(S data)? listener, {
+    bool disposeOnDone = false,
+  }) {
     return subscribe(Computable.fromFuture<S?>(future), (value) {
       listener?.call(value);
     });
   }
 
-  /// Subscribes the [Computable] to the provided source computable sstream and forwards the values it emits
+  /// Subscribes the [Computable] to the provided source computable stream and forwards the values it emits
   /// onto the [Computable].
   StreamSubscription<T> forward(
     Computable<T?> computable, {
-    bool disposeOnDone = true,
+    bool disposeOnDone = false,
   }) {
     StreamSubscription<T>? subscription;
-    subscription = subscribe(computable, add)
-      ..onDone(() {
-        subscription!.cancel();
-
-        if (disposeOnDone) {
-          dispose();
-        }
-      });
+    subscription = subscribe(computable, add, disposeOnDone: disposeOnDone);
     _subscriptions.add(subscription);
     return subscription;
   }
@@ -82,10 +89,13 @@ class ComputableSubscriber<T> with ComputableMixin<T> implements Computable<T> {
     Stream<T> stream, {
     bool disposeOnDone = true,
   }) {
-    return forward(Computable.fromStream<T>(stream, initialValue: get()));
+    return subscribeStream(stream, add, disposeOnDone: disposeOnDone);
   }
 
-  StreamSubscription<T> forwardFuture(Future<T> future) {
-    return forward(Computable.fromFuture<T>(future, initialValue: get()));
+  StreamSubscription<T> forwardFuture(
+    Future<T> future, {
+    bool disposeOnDone = true,
+  }) {
+    return subscribeFuture(future, add, disposeOnDone: disposeOnDone);
   }
 }