+++ lib/computables.dart
@@ -2,13 +2,14 @@ library computables;
 
 import 'dart:async';
 
-export 'widgets/computable_stream_builder.dart';
+export 'widgets/computable_builder.dart';
 
 part 'computable_future.dart';
 part 'computable_stream.dart';
 part 'computation.dart';
 part 'computable_mixin.dart';
 part 'computable.dart';
+part 'computation_transform.dart';
 part 'computable_subscriber.dart';
 part 'extensions/stream/start_with.dart';
 part 'extensions/stream/transformer.dart';