+++ test/computables_test.dart
@@ -151,6 +151,67 @@ void main() {
     });
   });
 
+  group("Computation transforms", () {
+    test('Subscribes to the output computable', () {
+      final computable = Computable(1);
+      final computable2 = Computable(2);
+
+      final computation = Computable.transform2(
+        computable,
+        computable2,
+        (input1, input2) {
+          final fill = input2 - input1;
+          return Computable.fromIterable(List.filled(fill, 0));
+        },
+      );
+
+      expectLater(
+        computation.stream(),
+        emitsInOrder([0]),
+      );
+    });
+
+    test(
+        'Updates the output computable when an input computable emits an updated value.',
+        () async {
+      final computable = Computable(1);
+      final computable2 = Computable(2);
+
+      final computation = Computable.transform2(
+        computable,
+        computable2,
+        (input1, input2) {
+          return Computable(input1 + input2);
+        },
+      );
+
+      computable2.add(5);
+
+      expectLater(computation.stream(), emitsInOrder([3, 6]));
+    });
+
+    test(
+        'Does not update the output computable when an input computable emits the same value',
+        () async {
+      final computable = Computable(1);
+      final computable2 = Computable(2);
+
+      final computation = Computable.transform2(
+        computable,
+        computable2,
+        (input1, input2) {
+          return Computable(input1 + input2);
+        },
+      );
+
+      // This computable update is ignored since its value is the same.
+      computable2.add(2);
+      computable2.add(3);
+
+      expectLater(computation.stream(), emitsInOrder([3, 4]));
+    });
+  });
+
   group('streamChanges', () {
     test("Streams value changes", () {
       final computable = Computable(2);