+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.4
+
+* Revamp builder
+
+## 0.0.3
+
+* Adds support for [Computable] transforms.
+
 ## 0.0.2
 
 * Core API changes.